+++ R/build_markdown.R
@@ -24,6 +24,8 @@ build_markdown <- function(path = ".", rebuild = FALSE, quiet = FALSE, slug = NU
   } else {
     create_site(path)
   }
+  # check if the lesson needs to be reset
+  this_lesson(path)
 
   episode_path <- path_episodes(path)
   outdir       <- path_built(path)
@@ -92,8 +94,10 @@ build_markdown <- function(path = ".", rebuild = FALSE, quiet = FALSE, slug = NU
     # Render the episode files to the built directory --------------------------
     renv_check_consent(path, quiet, sources)
     build_me <- db$build[needs_building]
+    slugs    <- get_slug(build_me)
 
     for (i in seq_along(build_me)) {
+      .html$set(slugs[i], "") # reset the html cache
       build_episode_md(
         path    = build_me[i],
         outdir  = outdir,