+++ R/utils-varnish.R
@@ -25,6 +25,7 @@ varnish_vars <- function() {
     sandpaper_version = ver("sandpaper"),
     sandpaper_cfg     = cfg("sandpaper"),
     pegboard_version  = ver("pegboard"),
+    pegboard_cfg      = cfg("pegboard"),
     varnish_version   = ver("varnish"),
     varnish_cfg       = cfg("varnish")
   )
@@ -51,12 +52,17 @@ varnish_vars <- function() {
 #'
 #' @keywords internal
 set_globals <- function(path) {
+  template_check$set()
   initialise_metadata(path)
   # get the resources if they exist (but do not destroy the global environment)
   old <- .resources$get()
   on.exit(.resources$set(key = NULL, old))
   set_resource_list(path)
   these_resources <- .resources$get()
+
+  for (slug in get_slug(unlist(these_resources, use.names = FALSE))) {
+    .html$set(slug, "")
+  }
 
   # Sidebar information is largely duplicated across the views. The only thing
   # that is different is the name of the index node.
@@ -78,6 +84,7 @@ set_globals <- function(path) {
 
   learner_globals$set(key = NULL, 
     c(list(
+      aio = TRUE,
       instructor = FALSE,
       sidebar = learner_sidebar,
       more = paste(learner$extras, collapse = ""),
@@ -86,6 +93,7 @@ set_globals <- function(path) {
   )
   instructor_globals$set(key = NULL, 
     c(list(
+      aio = TRUE,
       instructor = TRUE,
       sidebar = instructor_sidebar,
       more = paste(instructor$extras, collapse = ""),