+++ NEWS.md
@@ -1,3 +1,21 @@
+# sandpaper 0.4.0
+
+NEW FEATURES
+------------
+
+* An all-in-one page is now available for lesson websites at `/aio.html` and
+  `instructor/aio.html`. 
+
+MISC
+----
+
+* Provisioning of the global lesson element cache (metadata, AST, and global
+  variables for varnish) is now all executed via `this_lesson()`, which is run
+  during `validate_lesson()`. This simplifies the setup a bit, and provides the
+  same method of cache invalidation (checking git outputs) for all of these
+  elements
+
+
 # sandpaper 0.3.6
 
 MISC